+++ usage/tests/fails/missing_custom.rs
@@ -1,0 +1,11 @@
+use cache_diff::CacheDiff;
+
+#[derive(CacheDiff)]
+struct MissingCustom {
+    #[cache_diff(ignore = "custom")]
+    i_am_a_custom_field: String,
+
+    normal: String,
+}
+
+fn main() {}